+++ R/SampleSize_HVNTID.R
@@ -40,7 +40,8 @@ sampleN.HVNTID <- function(alpha=0.05, targetpower=0.8, theta0, theta1, theta2,
     dfRRe <- parse(text="n-2", srcfile=NULL)
     dfTTe <- parse(text="n-2", srcfile=NULL)
     # expectation of mse of the ANOVA of intra-subject contrasts T-R
-    Emse  <- (s2wT + s2wR)/2 
+    Emse  <- (s2wT + s2wR)/2
+    desi  <- "2x2x4 (TRTR|RTRT)"
   }
   if (design=="2x2x3") {
     seqs  <- 2
@@ -50,6 +51,7 @@ sampleN.HVNTID <- function(alpha=0.05, targetpower=0.8, theta0, theta1, theta2,
     dfTTe <- parse(text="n/2-1", srcfile=NULL)  # balanced only
     # expectation of mse of the ANOVA of intra-subject contrasts T-R
     Emse  <- 1.5*(s2wT + s2wR)/2                # balanced only
+    desi  <- "2x2x3 (TRT|RTR)"
   }
   
   mlog <- log(theta0)
@@ -62,7 +64,7 @@ sampleN.HVNTID <- function(alpha=0.05, targetpower=0.8, theta0, theta1, theta2,
     cat("\n+++++++++ FDA method for HV NTIDs ++++++++++++\n")
     cat("           Sample size estimation\n")
     cat("----------------------------------------------\n")
-    cat("Study design: ",design,"\n")
+    cat("Study design:",desi,"\n")
     cat("log-transformed data (multiplicative model)\n")
     cat(nsims,"studies for each step simulated.\n\n")
     cat("alpha  = ",alpha,", target power = ", targetpower,"\n", sep="")